#254: update other contracts, replace market with dispenser, remove old methods, update test.
