+++ squid_py/keeper/__init__.py
@@ -3,7 +3,7 @@
 from .contract_base import ContractBase
 from .didregistry import DIDRegistry
 from .keeper import Keeper
-from .market import Market
+from .dispenser import Dispenser
 from .service_execution_agreement import ServiceAgreement
 from .token import Token
 from .utils import (
